Audio Driver breakup (fist stage)
- moved out agc from main audio_driver.c
- removed OBSOLETE_NR and OBSOLETE_AGC code (which has not been in use
  for a while and has been replaced)
- moved tx code out of audio_driver.c into tx_processor.c
- worked on making tx functions less directly accessing the global buffers
 (instead getting them as parameters)
- simplified sideband & freq shift use. Now each modulator generates correct iq
  which can be passed through the final iq processing without IQ swap.

User improvements:
- fixed an issue with the twinpeaks calling Codec_RestartI2S in the interrupt handler
- fixed BPSK transmit to correctly transmit at selected sideband
- BPSK can now be entered using a cw paddle (like RTTY).
- RTTY now generates a sidetone on mcHF
- fixed wrong state buffer allocation for FreeDV RX IIR filter

Please test all tx modes. RX should be unaffected, no significant changes made in this area.
