+++ mchf-eclipse/drivers/audio/audio_management.c
@@ -414,91 +414,3 @@ void AudioManagement_SetSidetoneForDemodMode(uint8_t dmod_mode, bool tune_mode)
 
     softdds_configRunIQ(tonefreq,ts.samp_rate,0);
 }
-
-#ifdef OBSOLETE_AGC
-//
-//
-//*----------------------------------------------------------------------------
-//* Function Name       : UiCalcAGCDecay
-//* Object              : Calculate Decay timing for AGC (RECEIVE!)
-//* Input Parameters    :
-//* Output Parameters   :
-//* Functions called    :
-//*----------------------------------------------------------------------------
-//
-
-void AudioManagement_CalcAGCDecay()
-{
-    switch (ts.agc_mode)
-    {
-    case AGC_SLOW:
-        ads.agc_decay = AGC_SLOW_DECAY;
-        break;
-    case AGC_FAST:
-        ads.agc_decay = AGC_FAST_DECAY;
-        break;
-    case AGC_CUSTOM:      // calculate custom AGC setting
-    {
-        ads.agc_decay = powf(10,-((((float32_t)ts.agc_custom_decay)+30.0)/10.0));
-    }
-    break;
-    default:
-        ads.agc_decay = AGC_MED_DECAY;
-    }
-}
-
-void AudioManagement_CalcAGCVals(void)
-{
-    float max_rf_gain = 1 + (ts.max_rf_gain <= MAX_RF_GAIN_MAX ? ts.max_rf_gain : MAX_RF_GAIN_DEFAULT);
-
-    ads.agc_knee = AGC_KNEE_REF * max_rf_gain;
-    ads.agc_val_max = AGC_VAL_MAX_REF / max_rf_gain;
-    ads.post_agc_gain = POST_AGC_GAIN_SCALING_REF / (float)(ts.max_rf_gain + 1);
-    // TODO: Why is here always ts.max_rf_gain used? Shouldn't it be max_rf_gain too?
-}
-//
-//*----------------------------------------------------------------------------
-//* Function Name       : UiCalcRFGain
-//* Object              : Calculate RF Gain internal value from user setting
-//* Input Parameters    :
-//* Output Parameters   :
-//* Functions called    :
-//*----------------------------------------------------------------------------
-//
-void AudioManagement_CalcRFGain(void)
-{
-    float tcalc;    // temporary value as "ads.agc_rf_gain" may be used during the calculation!
-
-    // calculate working RF gain value
-    tcalc = (float)ts.rf_gain;
-    tcalc *= 1.4;
-    tcalc -= 20;
-    tcalc /= 10;
-    ads.agc_rf_gain = powf(10, tcalc);
-
-}
-//
-//
-//*----------------------------------------------------------------------------
-//* Function Name       : AudioManagement_CalcNB_AGC
-//* Object              : Calculate Noise Blanker AGC settings
-//* Input Parameters    :
-//* Output Parameters   :
-//* Functions called    :
-//*----------------------------------------------------------------------------
-
-void AudioManagement_CalcNB_AGC(void)
-{
-    float temp_float;
-
-    temp_float = ts.nb_agc_time_const;              // get user setting
-    temp_float = NB_MAX_AGC_SETTING-temp_float;     // invert (0 = minimum))
-    temp_float /= 1.1;                              // scale calculation
-    temp_float *= temp_float;                       // square value
-    temp_float += 1;                                // offset by one
-    temp_float /= 44000;                            // rescale
-    temp_float += 1;                                // prevent negative log result
-    ads.nb_sig_filt = log10f(temp_float);           // de-linearize and save in "new signal" contribution parameter
-    ads.nb_agc_filt = 1 - ads.nb_sig_filt;          // calculate parameter for recyling "old" AGC value
-}
-#endif
