+++ mchf-eclipse/drivers/ui/radio_management.h
@@ -332,6 +332,7 @@ void RadioManagement_Request_TxOn();
 void RadioManagement_Request_TxOff();
 
 bool RadioManagement_SwitchTxRx_Possible();
+bool RadioManagement_IsTxAtZeroIF(uint8_t dmod_mode, uint8_t digital_mode);
 
 inline void RadioManagement_ToggleVfoMem()
 {
